webapi/webapi-messaging-w3c-tests/messaging/MmsManager.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: MmsManager</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#mmsmanager-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var t1 = async_test("Check if the send method of mmsManager can send a MMS message successfully", {timeout: 2000});
var t2 = async_test("Check if the fetch method requests to fetch an MMS message with identifier equal to the indicated in the messageID parameter", {timeout: 2000});
var t3 = async_test("Check if the clear method of mmsManager delete all the messages associated to the messaging service", {timeout: 2000});

test(function () {
  assert_true(!!message, "The define of message");
  assert_equals(message.mms.type, "mms", "The type of mmsManager");
}, "Check if the type attribute is the type of the messaging service manager");

test(function () {
  assert_true(!!message, "The define of message");
  assert_equals(typeof(message.mms.serviceIDs), "string", "The serviceIDs of mmsManager");
}, "Check if the serviceIDs attribute is the identifier of the different services for this type of messaging service");

test(function () {
  assert_true(!!message, "The define of message");
  var fetchmode = message.mms.getFetchMode();
  assert_true(!!fetchmode, "GetFetchMode method requests to retrieve the fetch mode associated to a specific service");
}, 'Check if the getFetchMode method requests to retrieve the fetch mode associated to a specific service');

test(function () {
  assert_true(!!message, "The define of message");
  message.mms.setFetchMode("automatic");
  var fetchmode = message.mms.getFetchMode();
  assert_equals(fetchmode, "automatic", "the return value should same as the first param of setFetchMode method");
}, 'Check if the setFetchMode method request to the messaging system to set the MMS message fetch mode for the service');

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.mms.send({subject: "time", to: "friend"}).then(function (data) {
    t1.step(function () {
      assert_true(!!data, "A object will be returned in order to notify the result of the request");
    });
    t1.done();
  }, function (error) {
    t1.step(function () {
      assert_unreached(error.message);
    });
    t1.done();
  });
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.mms.fetch("0123456").then(function (data) {
    t2.step(function () {
      assert_true(!!data, "A object will be returned in order to notify the result of the request");
    });
    t2.done();
  }, function (error) {
    t2.step(function () {
      assert_unreached(error.message);
    });
    t2.done();
  });
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.mms.clear().then(function (data) {
    t3.step(function () {
      assert_true(!!data, "A object will be returned in order to notify the result of the request");
    });
    t3.done();
  }, function (error) {
    t3.step(function () {
      assert_unreached(error.message);
    });
    t3.done();
  });
});

</script>
